Comments on leaderboards
Ts is actaully aids. We save the comments indevidually to each roast, we save it with user id and name. Can only comment when logedin.

Goes ttrough the profanity filter then gets popsted.

Changed the db pulls dor leaderboard to allow the count of comments to be present

Server kinda cooked ngl.

autenticateToken is a lost cause.

Components are credited to chatgpt n cursor. Godbless